Renaming api to utils
