+++ pkg/bot/telegram.go
@@ -9,12 +9,15 @@ import (
 
 	botsecrets "github.com/julwrites/BotSecrets"
 
-	"github.com/julwrites/ScriptureBot/pkg/api"
+	"github.com/julwrites/ScriptureBot/pkg/utils"
+
+	"github.com/julwrites/BotMultiplexer/pkg/def"
+	"github.com/julwrites/BotMultiplexer/pkg/platform"
 )
 
 func TelegramHandler(res http.ResponseWriter, req *http.Request, secrets *botsecrets.SecretsData) {
 	log.Printf("Loading session data...")
-	env, ok := def.TranslateToProps(req, def.TYPE_TELEGRAM)
+	env, ok := platform.TranslateToProps(req, def.TYPE_TELEGRAM)
 	if !ok {
 		log.Printf("This message was not translatable to bot language")
 		return
@@ -23,16 +26,16 @@ func TelegramHandler(res http.ResponseWriter, req *http.Request, secrets *botsec
 	env.Secrets = *secrets
 	log.Printf("Loaded secrets...")
 
-	env = api.RegisterUser(env.User, env.Secrets.PROJECT_ID)
+	env.User = utils.RegisterUser(env.User, env.Secrets.PROJECT_ID)
 	log.Printf("Loaded user...")
 
 	env = HandleBotLogic(env)
 	log.Printf("Handled bot logic...")
 
-	if !def.PostFromProps(env) {
+	if !platform.PostFromProps(env) {
 		log.Printf("This message was not translatable from bot language")
 		return
 	}
 
-	api.PushUser(env.User, env.Secrets.PROJECT_ID) // Any change to the user throughout the commands should be put to database
+	utils.PushUser(env.User, env.Secrets.PROJECT_ID) // Any change to the user throughout the commands should be put to database
 }
